Adds the GUI with basic functionalities
A GUI for the application has been added with basic functionality.
This main of the GUI is configured to receive a video as parameter,
which will be automatically reproduced.

Although the GUI already includes several menus and controls in the
tool and status bars, they are only placeholders whose functionality
should be added in the future.