Converts detection actions into a toggle action
The actions for starting and stopping object detection have been
combined into a single toggle action. To do so, a "detect object"
property has been added to the display status.

As part of this change, the responsibility to enable/disable object
detection has been moved to the DefaultVideoPanelBuilder class,
although it could be placed in any other class, it was the class
most related to the functionality.

Finally, as part of this commit, the management of the fullscreen and
resizable actions of the menu has been unified and now it is the same
way as the detect object action.