Fixes Qt event management using signals
The management of the events manipulating the Qt components has been
fixed, as it was currently done modifying the components directly
from non-Qt threads (which is incorrect and causes failures).

As part of this change, the management of some events (such as the
video progress, threshold control, etc.) has been completed.