Simplifies event management
Event management in Qt components have been simplified with two main
changes: the GuiBuilder custom thread has been removed and the parent
components now are built by the builders.

The custom GuiBuilder thread was really not useful and, therefore, it
was removed. Instead, the VideoPanel widget now has its own thread to
do the image conversion. This avoid blocking the frame processor
thread while the conversion is done and, on the other hand, it avoids
blocking the Qt's event loop.

By moving the parent component creation to the builders, more control
is given to them and, in addition, the component is associated with
the builder, as it is a QObject.