Modifies the management of input devices
Creates some new classes that will help on the management of input devices. Now
the devices are listed when the device chooser dialog is opened, instead of
listing them on the startup of the application.
Also the devices are now represented by a class, that improves the behaviour,
allowing to set input using different OpenCV backends as GStreamer using string
instead of integer as access point. This devices are listed by a DeviceProvider
that checks which devices are connected.
This upgrades are introduced on the MenuBarBuilder and on the GuiArgumentParser
to allow changing this provider for a specific version.