Adds video recording functionality
Adds video recording action to the GUI. This function is done by selecting or
creating the file where the video is going to be recorded. This file could be
in AVI or MP4 format. Once it's selected, recording can start using the button.
If the name given to create a file does not have a extension, by default it
will be AVI.
Also the Open File action it's modified to accept AVI format videos.