Adds shortcuts configuration dialog and saving it to XML
Adds a new dialog, launched from the menu bar option "Configuration" and
"Shortcuts", used for configurate application shortcuts. This dialog is created
dinamically on the startup based on the shortcuts defined. It's organized in
tabs based on the shortcut's assigned group. It allows the user to change the
keys used to fire an action, and also default values could be restored.
Also a saving system is added, saving shortcuts changes every time them are
applied, to an XML file, that also is readed on the applicattion startup to
setup the system as it was  before closing.