Adds analyzer configuration dialog
Addss a new window designed for selecting and configuring the live analyzer for
the system. This window is created dinamically based on the analyzer properties
that defines the attributes it has. A new factory has the responsability of
create the appropiate widget for each data type.
Also inside the properties it's possible to define hints to add configuration
to the widget, as number range o step, and if the attribute may be hidden or
not.
Also some forgotten comments are deleted.