Fixes some errors
Makes Shortcuts configuration dialog and Analyzer configuration dialog as child
to the main window, to improve windows behaviour.
Changes Shortcut storage file name to make it clear with the last added
classes.
Fix a bug on analyzer configuration that makes that sucessive changes on the
same analyzer overwrite the firsts done.