Adds live analyzer configuration saving and new saving mode
Adds live analyzer configuration save and load to the same XML file used
previously for the shortcuts configuration. It will be saved with each change
to the analyzer and loaded on the application startup.
Also now the file writing is safer. It's saved to a temporal ".xml.new" file,
then the older file is overwritten whit the new version, and then the temporal
file is removed.