Adds support to live analysis activation in status
FrameProcessingStatus now has direct support for activating and
deactivating live analysis. This was currently done by externally
changing the analyzer to NoFrameChange.

Moving this responsibility to the status makes it more reusable and
avoids some programming errors.

This change also implies the elimination of the "detect_object"
property in the DisplayStatus class, as it has become redundant.

As part of this change, the ConfigurationStorage and the
configuration management have been improved. Now, the application
supports specifying the configuration file to load, starting with
default configuration and disabling automatic configuration
storage.