Adds configuration profiles for live analyzer
Adds support for creating and managing configuration profiles for live analyzer
This includes a new dialog for create a new configuration and give it a name
and select the model to be used. Once this is done, the parameters modification
form is showed to let changes on it before creating and applying it. On this
form, the old combo box that showed models, now shows cofigurations created,
and let to select one and activate it.
Also the storage system is updated to read and save this profiles from
configuration file.
Last, a little bug on color choose widget that makes the opening button dosn't
show the choosen color is fixed.