Modifies analyzer configuration selection dialog view
Modifies the dialog view, with a new form wrapping group box layout, where the
title shows the model selected name. Inside this layout there are two new
buttons for saving and removing a configuration. This way, saving and
activating a configuration are two actions with different buttons assigned.
Also Frame Processing Status is modified, including now methods for setting
the list of configurations and the active configuration index.
These methods derives into the creation of two new events to be emited when the
configurations list or the active index are modified. Also the responsability
of building the analyzer is moved from the dialog to the frame processing
status.