Updates analyzer configuration dialog to generalize it
Makes the analyzer configuration dialog works based on configurations list and
active configuration index received as parameters, and emiting signals when
there are changes saved or applied. This signals are configured with two
callback methods received as parameters as well, refering to two methods on
menu bar builder that act as intermediaries with the frame status.
Also the configuration object is now called FrameAnalyzerConfiguration instead
the old LiveAnalyzerConfiguration to made it reusable for static analyzers too.