Adds configuration dialog and storage for static analyzer
Adds support for configurating and storing configurations profiles for static
or image analyzer. The workflow is the same as for the video analyzers, as well
as the dialog and storage system, with little differences related to distinct
way of working between them. On the configuration file there is a new label
called static-analyzer-configuration for this.