Adds support for hiding the manu and tool bars on fullscreen
Upgraeds the fullscreen view to always hide the menu bar when activated. Also
there is a new option on the menu to activate the option to hide all the other
elemnts on the screen as the video controls bar or the status bar. This option
can be activated throw the menu or with a shortcut.
The shortcuts for activate and disactivate fullscrenn view and hiding controls
are also usable on the fullscreen view. Additionaly, fullscreen view can be
exited using 'Esc' key.
For supporting this, new methods on the display status and a new event were
created, as well as some modifications to the BorderLayout to allow the
complete hiding of the bars.